fix: rename directory typo hander → handler
Fixes #65

- Rename internal/platform/http/hander to internal/platform/http/handler
- Update import path in router.go
