+++ internal/app/router/router.go
@@ -4,7 +4,7 @@ import (
 	authhandler "stock_backend/internal/feature/auth/transport/handler"
 	candleshandler "stock_backend/internal/feature/candles/transport/handler"
 	symbollisthandler "stock_backend/internal/feature/symbollist/transport/handler"
-	handler "stock_backend/internal/platform/http/hander"
+	handler "stock_backend/internal/platform/http/handler"
 	jwtmw "stock_backend/internal/platform/jwt"
 
 	"github.com/gin-gonic/gin"
