+++ internal/platform/http/handler/health_handler.go
